# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/hello/{name}":{"get":{"summary":"Say Hello","operationId":"say_hello_hello__name__get","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string","title":"Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tasks/add_tasks/":{"post":{"tags":["tasks"],"summary":"Create Task","description":"Creation d'une nouvelle tache","operationId":"create_task_tasks_add_tasks__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tasks/tasks/priority_queue":{"get":{"tags":["tasks"],"summary":"Get Priority Queue","description":"Récupère les tâches organisées par date la plus proche","operationId":"get_priority_queue_tasks_tasks_priority_queue_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Task"},"type":"array","title":"Response Get Priority Queue Tasks Tasks Priority Queue Get"}}}}}}},"/tasks/tasks/priority":{"get":{"tags":["tasks"],"summary":"Get Priority Tasks","description":"Récupère les tâches organisées par ordre de priorité.","operationId":"get_priority_tasks_tasks_tasks_priority_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Task"},"type":"array","title":"Response Get Priority Tasks Tasks Tasks Priority Get"}}}}}}},"/tasks/get_tasks/":{"get":{"tags":["tasks"],"summary":"Get Tasks","description":"Recuperer la liste des tasks","operationId":"get_tasks_tasks_get_tasks__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/tasks/tasks/{task_id}":{"put":{"tags":["tasks"],"summary":"Update Task","description":"Mise à jour d'une tâche","operationId":"update_task_tasks_tasks__task_id__put","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"integer","title":"Task Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tasks/del_tasks/{task_id}":{"delete":{"tags":["tasks"],"summary":"Delete Task","description":"Supprime une tâche par son ID.","operationId":"delete_task_tasks_del_tasks__task_id__delete","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"integer","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/categories/categories/":{"get":{"tags":["categories"],"summary":"Get Categories","operationId":"get_categories_categories_categories__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Category-Output"},"type":"array","title":"Response Get Categories Categories Categories  Get"}}}}}},"post":{"tags":["categories"],"summary":"Create Category","description":"Crée une nouvelle catégorie et la stocke dans la liste des catégories.\nArgs: category (Category): La catégorie à créer.\nReturns:Category: La catégorie créée.\nRaises:ValueError: Si une catégorie avec le même nom existe déjà.","operationId":"create_category_categories_categories__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Category-Input"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Category-Input":{"properties":{"category_id":{"type":"integer","title":"Category Id"},"name":{"type":"string","title":"Name"},"tasks":{"items":{"$ref":"#/components/schemas/Task"},"type":"array","title":"Tasks"},"subcategories":{"items":{"$ref":"#/components/schemas/Category-Input"},"type":"array","title":"Subcategories"}},"type":"object","required":["category_id","name","tasks","subcategories"],"title":"Category"},"Category-Output":{"properties":{"category_id":{"type":"integer","title":"Category Id"},"name":{"type":"string","title":"Name"},"tasks":{"items":{"$ref":"#/components/schemas/Task"},"type":"array","title":"Tasks"},"subcategories":{"items":{"$ref":"#/components/schemas/Category-Output"},"type":"array","title":"Subcategories"}},"type":"object","required":["category_id","name","tasks","subcategories"],"title":"Category"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Task":{"properties":{"task_id":{"type":"integer","title":"Task Id"},"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"due_date":{"type":"string","format":"date","title":"Due Date"},"priority":{"type":"integer","title":"Priority"},"category":{"type":"string","title":"Category"}},"type":"object","required":["task_id","title","description","due_date","priority","category"],"title":"Task"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}